rename waterAllocation to useWaterRightQuery
